+++ mocks/admin-api-client-unit-mocks/mocks/collections.js
@@ -1,10 +1,10 @@
 module.exports = [
   {
     id: "base",
-    routesVariants: ["get-user:1"],
+    routes: ["get-user:1"],
   },
   {
     id: "user2",
-    routesVariants: ["get-user:2"],
+    routes: ["get-user:2"],
   },
 ];