+++ packages/plugin-admin-api/test/fixtures/files-error-routes/collections.js
@@ -11,16 +11,16 @@ Unless required by applicable law or agreed to in writing, software distributed 
 module.exports = [
   {
     id: "base",
-    routesVariants: ["get-users:success", "get-user:1"],
+    routes: ["get-users:success", "get-user:1"],
   },
   {
     id: "user-2",
     from: "base",
-    routesVariants: ["get-user:2"],
+    routes: ["get-user:2"],
   },
   {
     id: "user-real",
     from: "base",
-    routesVariants: ["get-user:real"],
+    routes: ["get-user:real"],
   },
 ];