+++ packages/plugin-proxy/test/fixtures/filters/collections.js
@@ -1,6 +1,6 @@
 module.exports = [
   {
     id: "base",
-    routesVariants: ["proxy-all:enabled", "user-2:default"],
+    routes: ["proxy-all:enabled", "user-2:default"],
   },
 ];