+++ packages/plugin-proxy/test/fixtures/host/collections.js
@@ -1,6 +1,6 @@
 module.exports = [
   {
     id: "base",
-    routesVariants: ["get-users:success", "get-user:real"],
+    routes: ["get-users:success", "get-user:real"],
   },
 ];