+++ test/core-e2e/src/fixtures/validations/collections.js
@@ -1,16 +1,16 @@
 module.exports = [
   {
     id: "base",
-    routesVariants: ["get-users-invalid:success", "get-user-variant-invalid:1", "get-user:2"],
+    routes: ["get-users-invalid:success", "get-user-variant-invalid:1", "get-user:2"],
   },
   {
     id: "base",
-    routesVariants: [],
+    routes: [],
   },
   {
     id: "invalid-variant",
     from: "base",
-    routesVariants: ["get-user-variant-invalid:2"],
+    routes: ["get-user-variant-invalid:2"],
   },
   {
     id: "invalid-mock",
@@ -18,11 +18,11 @@ module.exports = [
   {
     id: "invalid-from",
     from: "foo",
-    routesVariants: ["get-user:2"],
+    routes: ["get-user:2"],
   },
   {
     id: "duplicated-route",
     from: "base",
-    routesVariants: ["get-user:2", "get-user:1"],
+    routes: ["get-user:2", "get-user:1"],
   },
 ];