+++ test/core-e2e/src/load-collections-and-routes.spec.js
@@ -122,17 +122,17 @@ describe("loadCollections and loadRoutes methods", () => {
         {
           id: "users-and-library",
           from: "base",
-          routesVariants: ["get-books:success", "get-authors:success"],
+          routes: ["get-books:success", "get-authors:success"],
         },
         {
           id: "authors-error",
           from: "users-and-library",
-          routesVariants: ["get-authors:error"],
+          routes: ["get-authors:error"],
         },
         {
           id: "bradbury",
           from: "users-and-library",
-          routesVariants: ["get-authors:bradbury"],
+          routes: ["get-authors:bradbury"],
         },
       ]);
       core.config.namespace("mock").namespace("collections").option("selected").value =