+++ test/main-e2e/src/fixtures/proxy/mocks/collections.js
@@ -1,10 +1,10 @@
 module.exports = [
   {
     id: "base",
-    routesVariants: ["proxy-all:enabled"],
+    routes: ["proxy-all:enabled"],
   },
   {
     id: "proxy-disabled",
-    routesVariants: ["proxy-all:disabled"],
+    routes: ["proxy-all:disabled"],
   },
 ];